server/requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==0.19.0
PyJWT==2.3.0
flask-cors
flask-restx
prometheus-client
psutil
flask-debugtoolbar 